Calculus

Description

Miscellaneous simple functions.

Documentation

absmin :: (Ord a, Num a) => a -> a -> a

($?) :: Bool -> Maybe a -> Maybe a

If False, turns a Maybe into Nothing.

eqs :: Eq b => (a -> b) -> a -> a -> Bool

Equality by applying a function to both arguments. Goes well with f2all, e.g. f2all [eqs recordFieldA, recordFieldB].

f2 :: a -> b -> [a -> b -> c] -> [c]

Applies two arguments to a list of functions that accept two arguments.

f2all :: [a -> a -> Bool] -> a -> a -> Bool

and f2. Goes well with f2all, e.g. f2all [eqs recordFieldA, recordFieldB].

tInit :: Text -> Text

Text init that returns "" if given "".

tTail :: Text -> Text

Text tail that returns "" if given "".

doIf :: Bool -> (a -> a) -> a -> a

Second argument if True, otherwise id.

(?) :: Bool -> (a -> a) -> a -> a infixr 8

doIf

out2 :: Bool -> (a, a) -> a

fst if True, snd if False.

in2 :: Bool -> b -> (b, b) -> (b, b)

Sets first if True, second if False.

do2 :: Bool -> (a -> a) -> (a, a) -> (a, a)

Applies a function to first if True, second if False.

both :: (a -> b) -> (a, a) -> (b, b)

Applies a function to both.

map2 :: (a -> b) -> [(a, a)] -> [(b, b)]

map . both.

duplic :: Eq a => [a] -> Bool

List contains duplicates.

tshow :: Show a => a -> Text

pack . show

seqConcat :: Foldable a => a (Seq b) -> Seq b

Seq concatenation.

seqMap :: (a -> Seq b) -> Seq a -> Seq b

seqConcat . <$>

seqMaybes :: Seq (Maybe a) -> Seq a

Seq equivalent of catMaybes.

tryJust :: Monad m => m a -> Maybe a -> m a

Extracts a Just in a monad. Returns the error message argument if given Nothing.

deleteOne :: (a -> Bool) -> [a] -> [a]

Deletes first element that matches a predicate.

pick :: StdGen -> [a] -> (a, StdGen)

Chooses an element of a list at random.

safePick :: StdGen -> [a] -> (Maybe a, StdGen)

Nothing if given an empty list, otherwise Just pick.

unfoldl :: (a -> Maybe (a, a)) -> a -> [a]

Left equivalent of unfoldr that ends when it reaches a Nothing.