Foundation

Description

Yesod foundation. Implemented in Application.

Documentation

data GameInfo

Constructors

GameInfo

Fields

gameVsWho :: Key User

gameVsUser :: User

gamePar :: Player

gameLeft :: Int

gameGame :: Game

newtype GameMessage

Constructors

Enact Game

data QueueMessage

Constructors

Announce (Key User) User [Character]

Respond (Key User) (TChan GameMessage) (TChan GameMessage) GameInfo

data App

The foundation datatype for your application. This can be a good place to keep settings and values requiring initialization before your application starts running, such as database connections. Every handler will have access to the data present here.

Constructors

App

Fields

appSettings :: AppSettings

appStatic :: Static

appConnPool :: ConnectionPool

appHttpManager :: Manager

appLogger :: Logger

appPractice :: Map (Key User) Game

appQueue :: TChan QueueMessage

data MenuItem

Constructors

MenuItem

Fields

menuItemLabel :: Text

menuItemRoute :: Route App

menuItemAccessCallback :: Bool

data MenuTypes

Constructors

NavbarLeft MenuItem

NavbarRight MenuItem

type Handler = HandlerFor App

type Widget = WidgetFor App ()

resourcesApp :: [ResourceTree String]

type DB a = forall (m :: * -> *). MonadIO m => ReaderT SqlBackend m a

A convenient synonym for database access functions.

newUser :: Text -> Maybe Text -> User

isAuthenticated :: Handler AuthResult

unsafeHandler :: App -> Handler a -> IO a