Game.Functions

Description

Functions for manipulating data in Game.Structure.

Text -> Slot arguments check for lMatch.

Arithmetic

incr :: Int -> Int

Adds 1 to positives, subtracts 1 from negatives, and leaves 0s unchanged.

byPar :: Int -> (a -> a) -> (a, a) -> (a, a)

Applies a function to the first in a pair if odd or the second if even.

opponent :: Player -> Player

Flips PlayerA and PlayerB.

lost

Arguments

:: Int
Interval

-> Int
Amount per interval

-> Int
Health lost

-> Int
Scaled amount

Scales an amount by health lost, in specified intervals.

illegal :: Player -> Act -> Bool

Chakras

(-~) :: Chakras -> Chakras -> Chakras infixl 6

Subtraction.

getGameChakra :: Player -> Game -> Chakras

setGameChakra :: Player -> Chakras -> Game -> Game

lack :: Chakras -> Bool

χ :: [ChakraType] -> Chakras

unχ :: Chakras -> [ChakraType]

χGain :: Player -> [ChakraType] -> Seq Ninja -> Chakras -> Chakras

Adds 1 chakra per living Ninja on the team of a Player.

χTotal :: Chakras -> Int

Sum of components.

Channel

fromChannel :: Ninja -> Channel -> Act

isChanneling :: Text -> Ninja -> Bool

Defense

hasDefense :: Text -> Slot -> Ninja -> Bool

Effect

filterEffects :: MonadPlus m => Ninja -> m Effect -> m Effect

Modifies Effects when they are first added to a Ninja.

Reduction of all Effects on a Ninja

getBoost :: Slot -> Ninja -> Int

Boost sum.

getBleed :: [Class] -> Ninja -> Int

Bleed sum.

getBless :: Ninja -> Int

Bless sum.

getBuild :: Ninja -> Int

Build sum.

getImmune :: Ninja -> [Class]

Immunes.

getLink :: Slot -> Ninja -> Int

Link sum.

getNet :: Player -> Game -> Ninja -> Int

Afflict sum minus Heal sum.

getReduce :: [Class] -> Ninja -> Int

Reduce sum.

getScale :: [Class] -> Ninja -> Rational

Scale product.

getShare :: Ninja -> [Slot]

Shares.

getSnare :: Ninja -> Int

Snare sum.

getStrengthen :: [Class] -> Ninja -> Int

Strengthen sum.

getStun :: Ninja -> [Class]

Stuns.

getWard :: [Class] -> Ninja -> Rational

Ward product.

getWeaken :: [Class] -> Ninja -> Int

Weaken sum.

Ninja

alter :: (Seq Ninja -> Seq Ninja) -> Game -> Game

Transforms all gameNinjas.

getCds :: Ninja -> Seq Int

Life and death

alives :: Player -> Seq Ninja -> [Ninja]

dead :: Player -> Game -> Bool

The entire team of a Player is dead, in which case they lose.

healthBound

Arguments

:: Int
Minimum (usually minHealth)

-> Int
Unbounded health

-> Int
Health between minimum and 100

Restricts nHealth within a range.

isAlive :: Ninja -> Bool

minHealth :: Ninja -> Int

1 if the Ninja is affected by Endure, otherwise 0.

Player

getPlayer :: Key User -> Game -> Maybe Player

getVs :: Key User -> Game -> Maybe (Key User)

yieldVictor :: Game -> Game

If a Game is over and gameVictor is unset, sets it to VictorA, VictorB, or Tie.

Skill

chakraClasses :: Skill -> Skill

Adds Bloodline, Genjutsu, Ninjutsu, Taijutsu, and Random to the classes of a Skill if they are included in its cost.

defaultL :: Text -> Skill -> Text

Replaces an empty string with a label.

Copied

copyDur :: Copying -> Int -> Int

Maximum duration of an effect. Effects from Copied Skills must not last longer than the copyDuration.

copyRoot :: Skill -> Slot -> Slot

Skill owner. Determines the folder location of the icon image.

Modification

getSkill :: Ninja -> Either Int Skill -> Skill

getSkill' :: Ninja -> Int -> Int -> Skill

Simplified getSkill that ignores Copied Skills and does not check if the skill is usable.

getSkills :: Ninja -> [Skill]

changes

(••) :: SkillTransform -> SkillTransform -> SkillTransform infixl 1

Composition.

addClass :: Class -> SkillTransform

changeWith :: Text -> SkillTransform -> SkillTransform

Applies a SkillTransform if hasOwn.

costPer :: Text -> [ChakraType] -> SkillTransform

Increases cost per numActive.

restrict :: SkillTransform

Turns AoE effects into single-target effects.

setCost :: [ChakraType] -> SkillTransform

swapSkill :: Status -> Skill -> Skill

Affects enemies instead of allies and allies instead of enemies.

targetAll :: SkillTransform

Turns single-target effects into AoE effects.

Requirement

matchRequire :: Requirement -> Slot -> Ninja -> Bool

Status

has :: Text -> Slot -> Ninja -> Bool

hasOwn :: Text -> Ninja -> Bool

is :: Effect -> Ninja -> Bool

is' :: (Class -> Effect) -> Ninja -> Bool

numActive :: Text -> Ninja -> Int

numStacks :: Text -> Slot -> Ninja -> Int

Trap

classTrs :: (Foldable a, Functor a) => Bool -> (Class -> Trigger) -> a Class -> Ninja -> Seq TrapTransform

getPerTraps :: Bool -> Trigger -> Int -> Ninja -> Seq (Game -> Game)

getTrackTraps :: Bool -> Trigger -> Ninja -> Seq (Game -> Game)

getTraps :: Bool -> Trigger -> Ninja -> Seq TrapTransform

Triggering Effects

copy :: [Class] -> Ninja -> Bool -> [(Slot, Text, Int, Int)]

Trigger a Copy. Returns (statusSrc, statusL, copyTo, copyDuration).

counter :: [Class] -> Ninja -> Ninja -> Maybe Ninja

Trigger a Counter.

parry :: Skill -> Ninja -> Maybe (Ninja, Slot, Int)

Trigger a Parry.

reapply :: [Class] -> Ninja -> Maybe Slot

Trigger a Reapply.

redirect :: [Class] -> Ninja -> Maybe Slot

Trigger a Redirect.

reflect :: [Class] -> Ninja -> Ninja -> Maybe Ninja

Trigger a Reflect.

snareTrap :: Skill -> Ninja -> Maybe (Ninja, Int)

Trigger a SnareTrap.

swap :: [Class] -> Ninja -> Maybe Status

Trigger a Swap.

Usability

targetable

Arguments

:: Skill

-> Ninja
Skill owner

-> Ninja
User

-> Ninja
Target

-> Bool

usable

Arguments

:: Ninja

-> Maybe Int
Index in characterSkills

-> Skill

-> Skill