Game.Game

Description

A collection of Transforms and other Game → Game transformations.

Documentation

setFace :: Int -> Transform

wait :: Transform

wait' :: a -> Transform

apply :: Int -> [Effect] -> Transform

apply' :: Text -> Int -> [Effect] -> Transform

applyWith :: [Class] -> Int -> [Effect] -> Transform

applyWith' :: [Class] -> Text -> Int -> [Effect] -> Transform

applyX :: Int -> (Int -> Effect) -> Int -> Transform

applyDur :: [Effect] -> Int -> Transform

applyDur' :: Text -> [Effect] -> Int -> Transform

prolong :: Int -> Text -> Transform

prolongChannel :: Int -> Text -> Transform

hasten :: Int -> Text -> Transform

refresh :: Text -> Transform

tag :: Int -> Transform

tag' :: Text -> Int -> Transform

hide :: Int -> [Effect] -> Transform

hide' :: Text -> Int -> [Effect] -> Transform

flag :: Transform

flag' :: Text -> Transform

bomb :: Int -> [Effect] -> [(Bomb, Transform)] -> Transform

bomb' :: Text -> Int -> [Effect] -> [(Bomb, Transform)] -> Transform

bombWith :: [Class] -> Int -> [Effect] -> [(Bomb, Transform)] -> Transform

bombWith' :: [Class] -> Text -> Int -> [Effect] -> [(Bomb, Transform)] -> Transform

snapshot :: Int -> Transform

remove :: Text -> Transform

cureStun :: Transform

cureBane :: Transform

cure :: (Effect -> Bool) -> Transform

cureAll :: Transform

purgeAll :: Transform

purge :: Transform

addStack :: Transform

addStacks :: Text -> Int -> Transform

addStacks' :: Int -> Text -> Int -> Transform

removeStack :: Text -> Transform

removeStacks :: Text -> Int -> Transform

setHealth :: Int -> Transform

heal :: Int -> Transform

restore :: Int -> Transform

copyLast :: Int -> Int -> Transform

copyAll :: Int -> Transform

teach :: Int -> (Slot -> Int -> Copying) -> Int -> Transform

teachOne :: Int -> Int -> (Slot -> Int -> Copying) -> Int -> Transform

sacrifice :: Int -> Int -> Transform

afflict :: Int -> Transform

pierce :: Int -> Transform

damage :: Int -> Transform

execute :: (Int -> Transform) -> Int -> Transform

leech :: Int -> (Int -> Transform) -> Transform

kill :: Transform

kill' :: Transform

killThen :: Transform -> Transform

defend :: Int -> Int -> Transform

defend' :: Text -> Int -> Int -> Transform

addDefense :: Text -> Int -> Transform

demolish :: Transform

demolish' :: Int -> Transform

commandeer :: Transform

bar :: Int -> (Int -> Transform) -> Transform -> Int -> Transform

factory :: Transform

resetAll :: Transform

reset :: Int -> Int -> Transform

alterCd :: Int -> Int -> Int -> Transform

resetCharges :: Transform

interrupt :: Transform

delay :: Int -> Transform -> Transform

trapWith :: TrapType -> [Class] -> Int -> Trigger -> Transform -> Transform

trap :: Int -> Trigger -> Transform -> Transform

trap' :: Int -> Trigger -> Transform -> Transform

trapFrom :: Int -> Trigger -> Transform -> Transform

trapFrom' :: Int -> Trigger -> Transform -> Transform

trapPer :: Int -> Trigger -> (Int -> Transform) -> Transform

trapPer' :: Int -> Trigger -> (Int -> Transform) -> Transform

onBreak :: Transform -> Transform

removeTrap :: Text -> Transform

gain :: [ChakraType] -> Transform

drain :: Int -> Transform

steal :: Int -> Transform

cancelChannel :: Text -> Transform

vary :: Int -> Int -> Int -> Transform

vary' :: Int -> Int -> Transform

self :: Transform -> Transform

Directly applies an effect to the original source without going through wrapEffect.

alliedTeam :: Transform -> Transform

Directly applies an effect to all allies of the original source without going through wrapEffect.

enemyTeam :: Transform -> Transform

Directly applies an effect to all enemies of the original source without going through wrapEffect.

everyone :: Transform -> Transform

Directly applies an effect to all characters without going through wrapEffect.

(°) :: Transform -> Transform -> Transform infixl 3

(•) :: Transform -> Transform -> Transform infixl 1

ifI :: Text -> Transform -> Transform

ifU :: Text -> Transform -> Transform

ifnotI :: Text -> Transform -> Transform

ifnotU :: Text -> Transform -> Transform

ifChan :: Text -> Transform -> Transform

ifHealthI :: Int -> Int -> Transform -> Transform

ifHealthU :: Int -> Int -> Transform -> Transform

ifStacks :: Text -> Int -> Transform -> Transform

ifnotStacks :: Text -> Int -> Transform -> Transform

ifDef :: Text -> Transform -> Transform

withInvulnU :: Int -> (Int -> Transform) -> Int -> Transform

withI :: Text -> Int -> (Int -> Transform) -> Int -> Transform

withU :: Text -> Int -> (Int -> Transform) -> Int -> Transform

withChan :: Text -> Int -> (Int -> Transform) -> Int -> Transform

perI :: Text -> Int -> (Int -> Transform) -> Int -> Transform

perI' :: Text -> Int -> Int -> (Int -> Transform) -> Int -> Transform

perU :: Text -> Int -> (Int -> Transform) -> Int -> Transform

perHealthI :: (Int -> Int) -> (Int -> Transform) -> Transform

perHealthU :: (Int -> Int) -> (Int -> Transform) -> Transform

perDef :: Text -> (Int -> Transform) -> Int -> Transform

perHelpful :: Int -> (Int -> Transform) -> Int -> Transform

perAffected :: Text -> Int -> (Int -> Transform) -> Int -> Transform

updateChakra :: Player -> [Int] -> Game -> Game

runTurn :: Player -> [Act] -> StdGen -> Game -> Game

forfeit

Arguments

:: Player

-> Game

-> Game
Sets all on team to 0 health.

censor :: Player -> Game -> Game

Hide private information and Invisibles from opponent.

invuln :: Text -> Text -> [Class] -> [Skill]

invuln' :: Text -> Text -> [Class] -> [Transform] -> [Skill]

act :: [Affected] -> Game -> (Act, Either (Slot, Slot) StdGen) -> Game

kabuto :: Transform