Game.Structure

Description

Data structures for gameplay.

Act

data Act

A single action of a Ninja.

Constructors

Act

Fields

actC :: Slot

User index in gameNinjas (0-5)

actS :: Either Int Skill

Skill by index in nCharacter characterSkills (0-3)

actT :: Slot

Target index in gameNinjas (0-5)

data ActPath

Builds an Act out of basic types. Used for PathPiece.

Constructors

ActPath

Fields

actPathC :: Int

actC

actPathS :: Int

Left actS

actPathT :: Int

actT

actFromPath :: ActPath -> Act

botActs :: [Act]

Actions of AI in training mode.

data Affected

Constructors

Applied

Channeled

Countered

Delayed

Disrupted

Parrying

Redirected

Reflected

Swapped

Trapped

Chakras

data Chakras

Collection of all chakra types.

Constructors

Chakras

Fields

blood :: Int

Bloodline

gen :: Int

Genjutsu

nin :: Int

Ninjutsu

tai :: Int

Taijutsu

rand :: Int

Random

data ChakraType

Types of chakra in Chakras.

Constructors

Blood
Bloodline

Gen
Genjutsu

Nin
Ninjutsu

Tai
Taijutsu

Rand
Random

Channel

data Channel

An Act channeled over multiple turns.

Constructors

Channel

Fields

channelRoot :: Slot

channelSkill :: Skill

channelT :: Slot

channelDur :: Channeling

data Channeling

Types of channeling for Skills.

Constructors

Instant

Passive

Action Int

Control Int

Ongoing Int

data ChannelTag

Indicates that a channeled Skill will affect a Ninja next turn.

Constructors

ChannelTag

Fields

tagRoot :: Slot

tagSrc :: Slot

tagSkill :: Skill

tagGhost :: Bool

tagDur :: Int

Character

data Character

An out-of-game character.

Constructors

Character

Fields

characterName :: Text

characterBio :: Text

characterSkills :: NonEmpty (NonEmpty Skill)

characterHooks :: [(Trigger, Ninja -> Int -> Ninja)]

Class

data Class

Qualifiers of Skills and Statuses.

Constructors

Invisible

InvisibleTraps

Soulbound

Bane

Summon

Melee

Ranged

Chakra

Physical

Mental

Nonstacking

Single

Multi

Extending

Bypassing

Uncounterable

Unreflectable

Unremovable

Necromancy

All

Hidden

Affliction

NonAffliction

NonMental

Shifted

Unshifted

Direct

BaseTrap

NewRandoms

Bloodline

Genjutsu

Ninjutsu

Taijutsu

Random

Delay

data Delay

Applies an effect after several turns.

Constructors

Delay

Fields

delayC :: Slot

delaySkill :: Skill

delayEf :: Game -> Game

delayDur :: Int

Destructible

data Barrier

Destructible barrier.

Constructors

Barrier

Fields

barrierAmount :: Int

barrierSrc :: Slot

barrierL :: Text

barrierWhile :: Game -> Game

barrierDone :: Int -> Game -> Game

barrierDur :: Int

data Defense

Destructible defense.

Constructors

Defense

Fields

defenseAmount :: Int

defenseSrc :: Slot

defenseL :: Text

defenseDur :: Int

Effect

data Effect

Effects of Statuses.

Constructors

Afflict Int
Deals damage every turn

AntiCounter
Cannot be countered or reflected

Bleed Class Int
Adds to damage received

Bless Int
Adds to healing Skills

Block
Treats source as Immune

Boost Int
Scales effects from allies

Build Int
Adds to destructible defense Skill

Counter Class
Counters the first Skills

CounterAll Class
Counters without being removed

Duel
Immune to everyone but source

Endure
Health cannot go below 1

Enrage
Ignore status effects

Exhaust Class
Skills cost an additional random chakra

Expose
Cannot reduce damage or be Immune

Focus
Immune to Stuns

Heal Int
Heals every turn

Immune Class
Invulnerable to enemy Skills

ImmuneSelf
Immune to self-caused damage

Isolate
Unable to affect others

Link Int
Increases damage and healing from source

Nullify Effect
Prevents effects from being applied

Parry Class Int
Counter and trigger a Skill

ParryAll Class Int
Parry repeatedly

Pierce
Damage skills turn into piercing

Plague
Immune to healing and curing

Reduce Class Int
Reduces damage by a flat amount

Reapply
Shares harmful skills with source

Redirect Class
Transfers harmful Skills

Reflect
Reflects the first Skill

ReflectAll
Reflect repeatedly

Restrict
Forces AoE attacks to be single-target

Reveal
Makes Invisible effects visible

Scale Class Rational
Scales damage dealt

Seal
Immune to friendly Skills

Share
Shares all harmful non-damage effects

Silence
Unable to cause non-damage effects

Snapshot Ninja
Saves a snapshot of the current state

Snare Int
Increases cooldowns

SnareTrap Class Int
Negates next skill and increases cooldown

Strengthen Class Int
Adds to all damage dealt

Stun Class
Unable to use Skills

Swap Class
Target swaps enemies and allies

Taunt
Forced to attack the source

Uncounter
Cannot counter or reflect

Unexhaust
Decreases chakra costs by 1 random

Unreduce Int
Reduces damage reduction Skills

Ward Class Rational
Reduces damage received by a fraction

Weaken Class Int
Lessens damage dealt

Copy
Copies a skill into source's skill slot

Fields

copyDuration :: Int

copyClass :: Class

copyTo :: Int

Skill index of source to copy into

copyNonHarm :: Bool

Include non-harmful Skills

helpful :: Effect -> Bool

sticky :: Effect -> Bool

Effect cannot be removed.

boost :: Int -> Effect -> Effect

Scales the power of an effect.

Function types

type Transform

Arguments

= Skill
Skill

-> Slot
Source (Src)

-> Slot
Actor (C)

-> Game
Before

-> Slot
Target (T)

-> Game
After

The type signature of game actions. Processed into Game → Game.

type SkillTransform = Ninja -> Skill -> Skill

The type signature of changes.

type TrapTransform

Arguments

= Int
Amount (optional argument for traps)

-> Slot
Source (optional argument for traps)

-> Game
Before

-> Game
After

The type signature of Trap actions.

Game

data Game

Game state.

Constructors

Game

Fields

gamePlayers :: (Key User, Key User)

gameNinjas :: Seq Ninja

gameChakra :: (Chakras, Chakras)

Starts at (Chakras 0 0 0 0 0, Chakras 0 0 0 0 0).

gameDelays :: [Delay]

Starts at [].

gameDrain :: (Int, Int)

Starts at (0, 0). Resets every turn to (0, 0).

gameSteal :: (Int, Int)

Starts at (0, 0). Resets every turn to (0, 0).

gameTraps :: Seq (Game -> Game)

Starts at empty.

gamePlaying :: Player

Starts at PlayerA.

gameVictor :: Maybe Victor

Starts at Nothing.

newGame :: [Character] -> Key User -> Key User -> Game

Constructs a Game with starting values.

gameNinja :: Slot -> Game -> Ninja

setNinja :: Slot -> Ninja -> Game -> Game

fn :: Slot -> (Ninja -> Ninja) -> Game -> Game

Labeled

class Labeled a where

Minimal complete definition

getL, getSrc

Methods

getL :: a -> Text

getSrc :: a -> Slot

lEq :: Labeled a => a -> a -> Bool

lMatch :: Labeled a => Text -> Slot -> a -> Bool

Ninja

data Ninja

In-game character, indexed between 0 and 5.

Constructors

Ninja

Fields

nId :: Slot

gameNinjas index (0-5)

nCharacter :: Character

nHealth :: Int

Starts at 100

nCooldowns :: Seq (Seq Int)

Starts at empty

nCharges :: Seq Int

Starts at 4 0s

nVariants :: Seq (NonEmpty Variant)

Starts at 4 0s

nCopied :: Seq (Maybe Copied)

Starts at 4 Nothings

nDefense :: [Defense]

Starts at []

nBarrier :: [Barrier]

Starts at []

nStatuses :: [Status]

Starts at []

nChannels :: [Channel]

Starts at []

newChans :: [Channel]

Starts at []

nTraps :: Seq Trap

Starts at empty

nFace :: [Face]

Starts at []

nParrying :: [Skill]

Starts at []

nTags :: [ChannelTag]

Starts at []

nLastSkill :: Maybe Skill

Starts at Nothing

data Face

Changes the character icon of a Ninja.

Constructors

Face

Fields

faceIcon :: Text

faceSrc :: Slot

faceDur :: Int

newNinja :: Character -> Slot -> Ninja

Constructs a Ninja with starting values from a character and an index.

ninjaReset :: Ninja -> Ninja

Factory resets a Ninja to its starting values.

adjustCd

Arguments

:: Int
Skill index (0-3).

-> Int
Variant index in characterSkills of nCharacter.

-> (Int -> Int)
Adjustment function.

-> Seq (Seq Int)

-> Seq (Seq Int)

insertCd

Arguments

:: Int
Skill index (0-3).

-> Int
Variant index in characterSkills of nCharacter.

-> Int
New cooldown.

-> Seq (Seq Int)

-> Seq (Seq Int)

Player

data Player

Player vs. opponent.

Constructors

PlayerA

PlayerB

data Victor

Constructors

VictorA

VictorB

Tie

Skill

data Skill

A move that a Character can perform.

Constructors

Skill

Fields

label :: Text

Name

desc :: Text

Description

require :: Requirement

Defaults to Usable

classes :: [Class]

Defaults to []

cost :: Chakras

Defaults to empty

cd :: Int

Defaults to 0

varicd :: Bool

Defaults to False

charges :: Int

Defaults to 0

channel :: Channeling

Defaults to Instant

start :: [(Target, Transform)]

Defaults to []

effects :: [(Target, Transform)]

Defaults to []

disrupt :: [(Target, Transform)]

Defaults to []

copying :: Copying

Defaults to NotCopied

skPic :: Bool

Defaults to False

changes :: Ninja -> Skill -> Skill

Defaults to id

newSkill :: Skill

Default values of a Skill. Used as a Skill constructor.

data Target

Target destinations of Skills.

Constructors

Self
User of Skill

Ally
Specific ally

Allies
All allies

RAlly
Random ally

XAlly
Specific ally excluding Self

XAllies
Allies excluding Self

Enemy
Specific enemy

Enemies
All enemies

REnemy
Random enemy

XEnemies
Enemies excluding Enemy

Everyone
All Ninjas

Specific Slot
Specific ninja index in gameNinjas (0-5)

data Requirement

Constructors

Usable

Unusable

HasI Int Text

HasU Text

data Copied

A Skill copied from a different character.

Constructors

Copied

Fields

copiedSkill :: Skill

copiedDur :: Int

Slot

data Slot

allied :: Slot -> Slot -> Bool

allies :: Slot -> Game -> [Ninja]

alliesP :: Player -> Seq Ninja -> [Ninja]

alliedP :: Player -> Slot -> Bool

enemies :: Slot -> Game -> [Ninja]

enemiesP :: Player -> Seq Ninja -> [Ninja]

[0 .. gameSize - 1].

spar :: Slot -> Int

teamSize :: Int

gameSize :: Int

allSlots :: [Slot]

allySlots :: Slot -> [Slot]

enemySlots :: Slot -> [Slot]

opponentSlots :: Player -> [Slot]

bySlot :: Slot -> (a -> a) -> (a, a) -> (a, a)

Obtains the first or second in a pair by parity.

outSlot :: Slot -> (a, a) -> a

Inverse of outSlot.

outSlot' :: Slot -> (a, a) -> a

Partition by parity.

choose :: (Maybe Slot, Maybe Slot) -> Target -> Slot -> Slot -> [Slot]

Translates a Target into a list of Ninjas.

skillTargets :: Skill -> Slot -> [Slot]

All targets that a Skill from a a specific Ninja affects.

Status

data Status

A status effect affecting a Ninja.

Constructors

Status

Fields

statusL :: Text

Label.

statusRoot :: Slot

Owner of the statusSkill

statusSrc :: Slot

Original user

statusC :: Slot

Direct user (e.g. if reflected)

statusSkill :: Skill

statusEfs :: [Effect]

statusClasses :: [Class]

statusBombs :: [(Bomb, Transform)]

statusMaxDur :: Int

statusDur :: Int

data Bomb

Applies Transforms when a Status ends.

Constructors

Done
Applied with both Expire and Remove

Expire
Applied when a Status reaches the end of its duration.

Remove
Applied when a Status is removed prematurely

data Copying

Constructors

Shallow Slot Int
No cooldown or chakra cost.

Deep Slot Int
Cooldown and chakra cost.

NotCopied

Trap

data Trap

A trap which gets triggered when a Ninja meets the conditions of a Trigger.

Constructors

Trap

Fields

trapType :: TrapType

trapTrigger :: Trigger

trapL :: Text

trapDesc :: Text

trapSrc :: Slot

trapEf :: TrapTransform

trapClasses :: [Class]

trapTrack :: Int

trapDur :: Int

data TrapType

Constructors

TrapTo

TrapFrom

TrapPer

data Trigger

Conditions to activate a Trap.

Constructors

OnAction Class

OnBreak Text

OnChakra

OnCounter Class

OnCounterAll

OnDamage

OnDamaged Class

OnDeath

OnHarm

OnHarmed Class

OnHealed

PerHealed

OnHelped

OnImmune

OnReflectAll

OnRes

OnStun

OnStunned Class

PerDamaged

TrackDamage

TrackDamaged

TurnBased

class TurnBased a where

Typeclass for structures that expire after a set number of turns.

Minimal complete definition

getDur, setDur

Methods

getDur :: a -> Int

Number of turns before expiration. If ≤ 0, never expires.

setDur :: Int -> a -> a

Updates the remaining number of turns after a turn has passed.

decrTurn :: TurnBased a => a -> Maybe a

sync :: Int -> Int

Variant

data Variant

Constructors

Variant

Fields

variantV :: Int

Index in characterSkills

variantVCD :: Bool

Uses a different cooldown than the baseline Skill

variantL :: Text

variantDur :: Int

variantCD :: Variant -> Int

noVariant :: Variant

Orphan instances

Eq (a -> b)

Methods

(==) :: (a -> b) -> (a -> b) -> Bool

(/=) :: (a -> b) -> (a -> b) -> Bool

ToJSON (a -> b)

Methods

toJSON :: (a -> b) -> Value

toEncoding :: (a -> b) -> Encoding

toJSONList :: [a -> b] -> Value

toEncodingList :: [a -> b] -> Encoding