Handler.Embed

Description

These handlers embed files in the executable at compile time to avoid runtime dependencies and improve efficiency.

Documentation

getFaviconR :: Handler TypedContent

getRobotsR :: Handler TypedContent