Calculus

Description

Miscellaneous simple functions.

Documentation

absmin :: (Ord a, Num a) => a -> a -> a

($?) :: Bool -> Maybe a -> Maybe a

If False, turns a Maybe into Nothing.

eqs :: Eq b => (a -> b) -> a -> a -> Bool

Equality by applying a function to both arguments. Goes well with andOn, e.g. andOn [eqs recordFieldA, recordFieldB].

enums :: (Bounded a, Enum a) => [a]

Lists all members of an Enum from minBound to maxBound.

andOn :: (Foldable a, Functor a) => a (b -> b -> Bool) -> b -> b -> Bool

Apply the same two arguments to a list of functions and and the result. Goes well with eq, e.g. andOn [eqs recordFieldA, recordFieldB].

tInit :: Text -> Text

Text init that returns "" if given "".

tTail :: Text -> Text

Text tail that returns "" if given "".

doIf :: Bool -> (a -> a) -> a -> a

Second argument if True, otherwise id.

(?) :: Bool -> (a -> a) -> a -> a infixr 8

doIf

out2 :: Bool -> (a, a) -> a

fst if True, snd if False.

in2 :: Bool -> b -> (b, b) -> (b, b)

Sets first if True, second if False.

do2 :: Bool -> (a -> a) -> (a, a) -> (a, a)

Applies a function to first if True, second if False.

both :: (a -> b) -> (a, a) -> (b, b)

Applies a function to both.

duplic :: Eq a => [a] -> Bool

List contains duplicates.

tshow :: Show a => a -> Text

pack . show

shorten :: Text -> Text

Removes spaces and special characters.

tryJust :: Monad m => m a -> Maybe a -> m a

Extracts a Just in a monad. Returns the error message argument if given Nothing.

deleteOne :: (a -> Bool) -> [a] -> [a]

Deletes first element that matches a predicate.

pick :: StdGen -> [a] -> (Maybe a, StdGen)

Chooses an element of a list at random.

unfoldl :: (a -> (a, a)) -> a -> NonEmpty a

Left equivalent of unfoldr that ends when it reaches a Nothing.