Game.Game

Description

Game handling, primarily through Transforms.

If a status Transform ends in an apostrophe, it uses a custom display name. If a trap Transform ends in an apostrophe, it is hidden from both players.

A common pattern in higher-order Transforms is (Int -> Transform) -> Int. This modifies the amount passed to a Transform that requires an amount, such as damage (how much?) or applyDur (how long?). (Int -> Transform) is the function, and the following Int is the base amount.

For example, withI "Example Status" 25 damage 50 results in "damage 50, and an additional 25 if the user has Example Status." This way, withI "Example Status" 25 looks like an adverb applied to damage 50, even though damage and 50 are actually two separate arguments of withI.

Handler interface

censor :: Player -> Game -> Game

Hides the opponent's team information and Invisible data.

forfeit :: Player -> Game -> Game

Kills all Ninjas of a Player, ending the game.

Turn handling

act :: [Affected] -> Game -> (Act, Either (Slot, Slot) StdGen) -> Game

Performs a single Act. If Right, this is a new action, in which case random targets are chosen by StdGen. If Left, this is a continuation of some previous action, in which case its random targets carry over from the previous action.

runTurn :: Player -> [Act] -> StdGen -> Game -> Game

The game engine's main function. Performs Acts and Channels; applies effects from Bombs, Barriers, Delays, and Traps; decrements all TurnBased data; adds ChannelTags; and resolves Chakras for the next turn.

updateChakra :: Player -> [Int] -> Game -> Game

Uses random integers to give a Player chakras at the start of their turn.

Higher-order Transform operations

Operators

() :: (a -> b) -> a -> b infixr 2

$

(•) :: Transform -> Transform -> Transform infixl 1

Function composition.

(°) :: Transform -> Transform -> Transform infixl 3

Same as '•', but with higher fixity.

Application

self :: Transform -> Transform

Directly applies an effect to the original source without checking anything (invulnerability, requirements, etc.)

alliedTeam :: Transform -> Transform

Directly applies an effect to all allies of the original source without checking anything (invulnerability, requirements, etc.)

enemyTeam :: Transform -> Transform

Directly applies an effect to all enemies of the original source without checking anything (invulnerability, requirements, etc.)

everyone :: Transform -> Transform

Directly applies an effect to all characters without checking anything (invulnerability, requirements, etc.)

Conditional

ifChan :: Text -> Transform -> Transform

User isChanneling

ifHealthI :: Int -> Int -> Transform -> Transform

User nHealth is within a range

ifHealthU :: Int -> Int -> Transform -> Transform

Target nHealth is within a range

ifI :: Text -> Transform -> Transform

User has

ifU :: Text -> Transform -> Transform

Target has

ifnotI :: Text -> Transform -> Transform

not ifI

ifnotU :: Text -> Transform -> Transform

not ifU

ifStacks :: Text -> Int -> Transform -> Transform

User numStacks exceeds a threshold

ifnotStacks :: Text -> Int -> Transform -> Transform

not ifStacks

killThen :: Transform -> Transform

Kills and performs a Transform if the target is killed. Has no effect if the target is invulnerable, has Endure, etc.

Scaling

perI :: Text -> Int -> (Int -> Transform) -> Int -> Transform

User numStacks

perI' :: Text -> Int -> Int -> (Int -> Transform) -> Int -> Transform

User numStacks divided by a denominator

perU :: Text -> Int -> (Int -> Transform) -> Int -> Transform

Target numStacks

perHealthI :: (Int -> Int) -> (Int -> Transform) -> Transform

User nHealth

perHealthU :: (Int -> Int) -> (Int -> Transform) -> Transform

Target nHealth

perAffected :: Text -> Int -> (Int -> Transform) -> Int -> Transform

Number of enemies matching has

perDef :: Text -> (Int -> Transform) -> Int -> Transform

Sum of user's matching Defense

perHelpful :: Int -> (Int -> Transform) -> Int -> Transform

Number of helpful Statuses and Defenses on the user that they did not cause themselves.

withI :: Text -> Int -> (Int -> Transform) -> Int -> Transform

User has

withU :: Text -> Int -> (Int -> Transform) -> Int -> Transform

Target has

withInvulnU :: Int -> (Int -> Transform) -> Int -> Transform

Target is Immune to any Class

withChan :: Text -> Int -> (Int -> Transform) -> Int -> Transform

User isChanneling

Transforms

delay :: Int -> Transform -> Transform

Adds a Delay.

factory :: Transform

Factory resets a Ninja with ninjaReset.

vary :: Int -> Int -> Int -> Transform

Adds a Variant.

vary' :: Int -> Int -> Transform

wait :: Transform

id.

wait' :: a -> Transform

const wait.

Chakras

drain :: Int -> Transform

gain :: [ChakraType] -> Transform

steal :: Int -> Transform

Channel

interrupt :: Transform

prolongChannel :: Int -> Text -> Transform

Combat

afflict :: Int -> Transform

Damage that cannot be reduced and penetrates Barrier and Defense.

damage :: Int -> Transform

Ordinary damage.

execute :: (Int -> Transform) -> Int -> Transform

Kills the target if their nHealth goes below a threshold.

pierce :: Int -> Transform

Damage that cannot be reduced.

heal :: Int -> Transform

Adds to nHealth.

leech :: Int -> (Int -> Transform) -> Transform

Steals nHealth from the target.

restore :: Int -> Transform

Restores a percentage of nHealth.

sacrifice :: Int -> Int -> Transform

Removes nHealth from the user.

Copying

copyAll :: Int -> Transform

copyLast :: Int -> Int -> Transform

Copies the target's nLastSkill.

teach :: Int -> (Slot -> Int -> Copying) -> Int -> Transform

Copies the user's Skills onto the target.

teachOne :: Int -> Int -> (Slot -> Int -> Copying) -> Int -> Transform

Copies a user's Skill onto the target.

Destructible

addDefense :: Text -> Int -> Transform

Adds to preexisting Defense.

bar :: Int -> (Int -> Transform) -> Transform -> Int -> Transform

Creates a Barrier.

defend :: Int -> Int -> Transform

Creates a Defense.

defend' :: Text -> Int -> Int -> Transform

demolish :: Transform

Removes the user's Barrier and the target's Defense.

demolish' :: Int -> Transform

Damage that only affects destructibles, not health.

Lifted

alterCd :: Int -> Int -> Int -> Transform

alterCd

cancelChannel :: Text -> Transform

cancelChannel

cure :: (Effect -> Bool) -> Transform

cure

cureAll :: Transform

cureBane :: Transform

cureBane

cureStun :: Transform

kill :: Transform

kill

kill' :: Transform

kill'

purge :: Transform

purge

refresh :: Text -> Transform

refresh

remove :: Text -> Transform

clear

removeStack :: Text -> Transform

removeStack

removeStacks :: Text -> Int -> Transform

removeStacks

reset :: Int -> Int -> Transform

reset

resetAll :: Transform

resetAll

resetCharges :: Transform

resetCharges

setHealth :: Int -> Transform

setHealth

Status

addStack :: Transform

addStacks :: Text -> Int -> Transform

addStacks' :: Int -> Text -> Int -> Transform

apply :: Int -> [Effect] -> Transform

Adds a Status to the target.

apply' :: Text -> Int -> [Effect] -> Transform

applyDur :: [Effect] -> Int -> Transform

applyDur' :: Text -> [Effect] -> Int -> Transform

applyX :: Int -> (Int -> Effect) -> Int -> Transform

bomb :: Int -> [Effect] -> [(Bomb, Transform)] -> Transform

bomb' :: Text -> Int -> [Effect] -> [(Bomb, Transform)] -> Transform

bombWith :: [Class] -> Int -> [Effect] -> [(Bomb, Transform)] -> Transform

flag :: Transform

flag' :: Text -> Transform

tag :: Int -> Transform

tag' :: Text -> Int -> Transform

hide :: Int -> [Effect] -> Transform

hide' :: Text -> Int -> [Effect] -> Transform

snapshot :: Int -> Transform

Modification

hasten :: Int -> Text -> Transform

hasten.

prolong :: Int -> Text -> Transform

prolong.

setFace :: Int -> Transform

Adds a Face with a duration.

Trap

onBreak :: Transform -> Transform

Applies an OnBreak trap for the Skill used.

removeTrap :: Text -> Transform

trap :: Int -> Trigger -> Transform -> Transform

Adds a TrapTo Trap.

trap' :: Int -> Trigger -> Transform -> Transform

trapFrom :: Int -> Trigger -> Transform -> Transform

Adds a TrapFrom Trap.

trapFrom' :: Int -> Trigger -> Transform -> Transform

trapPer :: Int -> Trigger -> (Int -> Transform) -> Transform

Adds a TrapPer Trap.

trapPer' :: Int -> Trigger -> (Int -> Transform) -> Transform

trapWith :: TrapType -> [Class] -> Int -> Trigger -> Transform -> Transform

Invulnerability

invuln :: Text -> Text -> [Class] -> NonEmpty Skill

invuln1 :: Text -> Text -> [Class] -> Skill

invuln' :: Text -> Text -> [Class] -> [Transform] -> NonEmpty Skill

Specialized

commandeer :: Transform

Steals all of the target's beneficial effects.

kabuto :: Transform